# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/songs": {
      "get": {
        "operationId": "SongsController_list",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Songs"
        ]
      },
      "post": {
        "operationId": "SongsController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDto"
                }
              }
            }
          }
        },
        "summary": "Upload an .mp3. Metadata comes from the file's ID3 tags; edit later via PATCH.",
        "tags": [
          "Songs"
        ]
      }
    },
    "/songs/{id}": {
      "get": {
        "operationId": "SongsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDto"
                }
              }
            }
          }
        },
        "tags": [
          "Songs"
        ]
      },
      "patch": {
        "operationId": "SongsController_retag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSongDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDto"
                }
              }
            }
          }
        },
        "tags": [
          "Songs"
        ]
      },
      "delete": {
        "operationId": "SongsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Songs"
        ]
      }
    },
    "/songs/{id}/audio": {
      "get": {
        "operationId": "SongsController_audio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full (200) or partial (206) audio stream",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "summary": "Audio bytes with HTTP Range support — `206` when a range is asked for (SNG-2, PB-1).",
        "tags": [
          "Songs"
        ]
      }
    },
    "/songs/{id}/cover": {
      "get": {
        "operationId": "SongsController_cover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cover image bytes",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "summary": "Cover image. Small and immutable for a given song, so cache it hard.",
        "tags": [
          "Songs"
        ]
      }
    },
    "/playlists": {
      "get": {
        "operationId": "PlaylistsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      },
      "post": {
        "operationId": "PlaylistsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      }
    },
    "/playlists/{id}": {
      "get": {
        "operationId": "PlaylistsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      },
      "patch": {
        "operationId": "PlaylistsController_rename",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenamePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      },
      "delete": {
        "operationId": "PlaylistsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Playlists"
        ]
      }
    },
    "/playlists/{id}/entries": {
      "post": {
        "operationId": "PlaylistsController_addEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      },
      "put": {
        "operationId": "PlaylistsController_reorder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderEntriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      }
    },
    "/playlists/{id}/entries/{entryId}": {
      "delete": {
        "operationId": "PlaylistsController_removeEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Playlists"
        ]
      }
    }
  },
  "info": {
    "title": "Web Music Player API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt"
        ]
      },
      "SongDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "hasCover": {
            "type": "boolean"
          },
          "addedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "ownerId",
          "title",
          "duration",
          "hasCover",
          "addedAt"
        ]
      },
      "UpdateSongDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "PlaylistEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "songId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "songId",
          "position"
        ]
      },
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistEntryDto"
            }
          },
          "trackCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "description": "When the playlist was created, as an ISO 8601 timestamp."
          }
        },
        "required": [
          "id",
          "ownerId",
          "name",
          "entries",
          "trackCount",
          "createdAt"
        ]
      },
      "CreatePlaylistDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "RenamePlaylistDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "AddEntryDto": {
        "type": "object",
        "properties": {
          "songId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "songId"
        ]
      },
      "ReorderEntriesDto": {
        "type": "object",
        "properties": {
          "entryIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "entryIds"
        ]
      }
    }
  }
}